unified_ai_agent/unified_mind/requirements.txt
# Unified Mind Dependencies

# Core dependencies
asyncio-compat>=0.1.0
dataclasses>=0.0  # Built-in, but explicit for clarity
typing>=0.0  # Built-in

# Logging
loguru>=0.7.0

# gRPC communication
grpcio>=1.60.0
grpcio-tools>=1.60.0

# Prometheus metrics
prometheus-client>=0.19.0

# LLM Providers (optional)
google-generativeai>=0.3.0  # Gemini API
llama-cpp-python>=0.2.0  # Local Llama (optional, requires compilation)

# NLP utilities
transformers>=4.36.0  # For embeddings and tokenization
sentence-transformers>=2.2.0  # For semantic search
torch>=2.1.0  # PyTorch for ML operations

# Async web server (for web interface)
aiohttp>=3.9.0
websockets>=12.0

# Configuration
pydantic>=2.5.0
python-dotenv>=1.0.0

# Utilities
rich>=13.7.0  # Rich console output
prompt-toolkit>=3.0.0  # Advanced CLI features

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
isort>=5.13.0
mypy>=1.8.0
